Added test: InnerText set at span in <dot:content>
