+++ src/DotVVM.Samples.Common/DotVVM.Samples.Common.csproj
@@ -20,6 +20,8 @@
     <None Remove="Views\ControlSamples\Button\ButtonOnclick.dothtml" />
     <None Remove="Views\ControlSamples\ComboBox\ComboBoxDelaySync.dothtml" />
     <None Remove="Views\ControlSamples\ComboBox\ComboBoxDelaySync2.dothtml" />
+    <None Remove="Views\ControlSamples\ContentPlaceHolder\DoubleContentPlaceHolderMasterPage.dotmaster" />
+    <None Remove="Views\ControlSamples\ContentPlaceHolder\DoubleContentPlaceHolderPage_ContentTest.dothtml" />
     <None Remove="Views\ControlSamples\IncludeInPageProperty\IncludeInPage.dothtml" />
     <None Remove="Views\ControlSamples\LinkButton\LinkButtonOnclick.dothtml" />
     <None Remove="Views\ControlSamples\Repeater\DataSourceNull.dothtml" />
@@ -66,6 +68,8 @@
     <Content Include="Scripts\testResource2.js" />
     <Content Include="tsconfig.json" />
     <Content Include="Views\ComplexSamples\CascadeSelectors\TripleComboBoxes.dothtml" />
+    <Content Include="Views\ControlSamples\ContentPlaceHolder\DoubleContentPlaceHolderMasterPage.dotmaster" />
+    <Content Include="Views\ControlSamples\ContentPlaceHolder\DoubleContentPlaceHolderPage_ContentTest.dothtml" />
     <Content Include="Views\ControlSamples\ValidationSummary\master.dotmaster" />
     <Content Include="Views\ControlSamples\ValidationSummary\IncludeErrorsFromTarget_PropertyPathNull.dothtml" />
     <Content Include="Views\ControlSamples\Button\ButtonOnclick.dothtml" />
