+++ src/DotVVM.Samples.Common/ViewModels/ControlSamples/ContentPlaceHolder/DoubleContentPlaceHolderMasterPageViewModel.cs
@@ -1,0 +1,14 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using DotVVM.Framework.ViewModel;
+
+namespace DotVVM.Samples.Common.ViewModels.ControlSamples.ContentPlaceHolder
+{
+    public class DoubleContentPlaceHolderMasterPageViewModel : DotvvmViewModelBase
+    {
+        
+    }
+}
+
